refactor(tools): split source.ts into per-surface/per-category modules (#692)
The 3,485-line packages/tools/src/registry/source.ts was the largest
non-generated file in the repo — one SourceTool interface plus three
module-private arrays (OVERLAP/AGENT_ONLY/MCP_ONLY) of 119 tool literals.
Any one-tool edit forced a large-diff review over a 105 KB file and the
three surface groups could not be reviewed or owned independently.

Move-only restructure (zero behavior change):
- SourceTool interface -> interfaces/source-tool.interface.ts (matches the
  interfaces/ convention used by tool-definition.interface.ts).
- registry/source/ replaces source.ts:
  - overlap.tools.ts (11 tools)
  - agent-only/ + mcp-only/ each split per category, with a barrel index.ts
    re-exporting AGENT_ONLY_TOOLS / MCP_ONLY_TOOLS
  - index.ts re-exports SOURCE_TOOLS = [...overlap, ...agentOnly, ...mcpOnly]
- tool-registry.ts import redirected source.js -> source/index.js (only change).
- Category placement follows tool-registry's own inferCategory() taxonomy.

Every tool literal was relocated verbatim (programmatic extraction, not
retyped); the resulting SOURCE_TOOLS deep-equals the pre-refactor export.

Adds a vitest config + contract test guarding the split: 119 tools, unique
names, frozen name inventory, surface partition (11/54/54), well-formed shape,
and a <=500-line budget per data module. No leaf module exceeds 500 lines.
