+++ packages/tools/src/interfaces/source-tool.interface.ts
@@ -1,0 +1,16 @@
+export interface SourceTool {
+  name: string;
+  description: string;
+  parameters: {
+    type: 'object';
+    properties: Record<string, unknown>;
+    required?: string[];
+  };
+  creditCost: number;
+  requiredRole: 'user' | 'admin' | 'superadmin';
+  surfaces: {
+    agent: boolean;
+    mcp: boolean;
+    cliAgentVisible?: boolean;
+  };
+}
